requirements.txt
certifi==2019.11.28
chardet==3.0.4
Django==2.2
django-rest-framework-mongoengine==3.4.1
djangorestframework==3.11.0
gcloud==0.17.0
googleapis-common-protos==1.51.0
httplib2==0.17.0
idna==2.8
influxdb==5.2.3
jws==0.1.3
mongoengine==0.19.1
oauth2client==3.0.0
pkg-resources==0.0.0
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodome==3.4.3
pymongo==3.10.1
Pyrebase==3.0.27
python-dateutil==2.8.1
python-jwt==2.0.1
pytz==2019.3
requests==2.22.0
requests-toolbelt==0.7.0
rsa==4.0
six==1.14.0
sqlparse==0.3.0
urllib3==1.25.8
